[FIX] account_asset: first depreciation amount
When an asset is set as Prorata Temporis, with a one month period, the
first depreciation amount is computed on the basis of the remaining days
in the purchase year. This doesn't make sense for a monthly
depreciation.

This commit makes the distinction between a monthly and a yearly
depreciation. In the case of a monthly depreciation, the remaining days
in the purchase month are taken into account.

opw-690034

Backport of bc2e825a205b1b0bdcba8b7143f106719c32efb3

The fiscal year start date should be taken into account when calculating
the depreciation amount in case of prorata temporis.

opw-678072

Backport of 5d8f74f7db7b5a1fb9a357e4e32a8a4e0b4307bd