[FIX] sale, website_portal_sale: see portal button
With 5e999303 the access action was refined depending on the user.

But when sending a mail about a sale order, this lead to an issue since
a mail template is rendered with the current user and not the mail
recipients.

So in the case of a sale order, the recipient could not have a button
leading to /my/orders/{order_id} whilst he should have, because the
user used to render the mail was the sender (which was an employee and
not a share user).

This commit adds a context key "force_website" which force the action to
an available /my/orders even if the sender is not a share user.

opw-691040