[IMP] website_forum: display 'comment' button under answers correctly
Fix up the commit https://github.com/odoo/odoo/commit/0a5870699a39fa074b50f3b08a3c09c9c3061252.
When no comment for an answer, the 'comment' button stays aligned with the others.
When there is a comment for an answer, the 'comment' button is under the last comment.