[IMP] website_event: put pager on bottom in website events page.
Purpose
========

When doing a filter on events page, the "second page" of the results is on the top right.
After I scroll down to the last event, nothing tells user that there are more events on the second page (or the second page is available)...

Specification
=============

Put 'Pager'(Pagination) on bottom right side in 'events page' so that user knows easilly to move on next page.

NB: Center the pager instead of displaying it on the right.