[FIX] web_editor: allow vimeo, youku and undefined video sources again
The refactoring made with https://github.com/odoo/odoo/commit/a1ffbaacab2b40cc89ee0a9b28b20082f7030ff8#diff-2b3acc31ced2b1e9f7522a3e29f33c22L825
broke the possibility to add vimeo, youku or other undefined video
sources.

This is because of jQuery. With jQuery you can create elements thanks to
multiple methods. From least efficient to most efficient:

- All in a string:
$("<tag attr1='value1' attr2='value2'/>");

- Tag in a string, attrs thanks to methods:
$("<tag/>").attr("attr1", "value1").attr("attr2", "value2");

- Tag in a string, attrs in second arg:
$("<tag/>", {attr1: "value1", attr2: "value2"});

The mentionned refactoring transformed the use of the second method
into the use of the last one. But there is a bug in jQuery: you cannot
use the first method and the last one all together. So:

$("<tag attr1='value1'/>", {attr2: "value2"});

will create a tag with attr1 but will totally ignore the attr2 part.

It happens that the allowfullscreen attr was left directly in the
creation string for vimeo and youku, preventing the src to be added.

This commit removes the use of allowfullscreen attr as fullscreen
is already handled by source players anyway.