[FIX] account: force to choose a refund sequence
On an accounting journal, when checking "Dedicated refund sequence",
choosing the refund sequence is mandatory. This is the subject of the
below revision:
a7ae991a73ff3f74091a55fb4f20ca2c84e68e42

Since the refund sequence field was available only for with the debug
mode, the user was not forced to choose the sequence, while he must be.

This is not possible to use the dedicated refund sequence if you do not
choose it there.

opw-693955