[FIX] sale_mrp: add flexibility to tests
test_sale_stock_lead_time.py compares dates and expect them to be the same.
Sometimes, the dates would differ by a second, because they are generated at
different moment during the execution.

This commit adds a tolerance of 1 second when comparing the dates.
(and changed from comparing Datetime strings to comparing timedeltas)