[FIX] base: unique bank account number constraint too restrictive
There is a multi-company record rule on `res.partner.bank`,
meaning you cannot see the bank accounts of other companies.

Making a constraint on `sanitized_acc_number` alone prevent
to create a same bank account into different companies,
while you cannot see this bank account from the other
companies.

In this case, the user tries to import its bank statement,
and it contains a bank account which already exists
in the database but for another company than the one
he is currently importing the statement, and,
 - it cannot find it because of the multi-company rule,
 - it cannot create it because of the constraint

The constraint must take into account the company

opw-690969