[FIX] stock: fix stock picking done quantity computation
This is an oversight during the port of
2524ccb2eb556ef5d17b845761ba85cebb70bd07
in the forward port revision
e364f2d5e2c06b9fbbee7acc4e3a8efac8ded1b4

In the revision 2524ccb2eb556ef5d17b845761ba85cebb70bd07,
the code has not only been moved,
but the variable passed to the method `rereserve_quants`
argument `move_ids` changed as well.
This has been missed during the port of the revision.

This leads to the invalid computation of the `done` quantity
in the pickings in some cases were the quants need to be
re-reserved.

opw-690244