Revert "[IMP] web_editor: don't lose modified code view content"
This reverts commit 19ca9b923f210568b2d8ca70195887f0cc33ea90.

It seems it breaks other use case, since this was fixing more a
usability thing (need to click back on code view to save code view
changes) it is better to revert than to add drawback.

A better fix should be introduced, or the solution would be to not fix
it and have a warning indicating what needs to be done.

opw-693052