[FIX] purchase: update order_lines date_planned
When we are trying to use the button 'set date to all order lines, we
get a traceback saying, that update cannot be done, because it ensure
that there's only one record. (After 9.0, there is no more this
'ensure_one').

This bug has been introduced in rev: https://github.com/odoo/odoo/commit/d583a57f843b711f2deac6f541847d1b74e65ea2

To fix it, we iterate over the order lines, and do the update of each
line.