[IMP] web_editor: allow lazy loadings in Animation instance
This commit comments the code of the Animation class and allows
for lazy loading of resources before animations' start methods
are called.

This lazy loading can be performed in the willStart method (to use
the same way as in the Widget class).

This commits also removes the "stop" function of the module as it
was not really useful. It was used to restart the animations (by
immediatly calling start again after the stop call). Now the Animation
class contains a restart method and the module "start" function
automatically knows if it has to instanciate an animation or restart
an existing one.
Animation start should only be performed by instanciating an Animation
(thanks to new) or by calling the restart method, to insure the
willStart method is always called. Anyway, the best way to start
animations is to let the system handle them, using the "selector"
attributes in your Animation definitions as before.

See in-code comments for details.