[FIX] purchase: double validation minimum amount
When we enable in the settings 'Get 2 levels of approvals' with a
'Double validation amount', and then create a request for quotation.
When we confirm it, even if the total amount of the request for
quotation is below the amount set in the settings, the purchase order
needs to be approved.

We excpect that if a purchase order has a total amount below the
validation amount setting, the purchase order is automatically
confirmed.

So to fix it, we compare the amount when the user click on the button
confirm. And immediatly call the function 'button_approve' if the amount
is below the double validation amount set in settings.

Bug introduced in rev: https://github.com/odoo/odoo/commit/cecc16b55

opw - 694381