[FIX] account, account_cancel: cancel payment
Commit ac555f8f613a prevents gaps in payment numbering by keeping tack
of the account move name used. It also prevents the unlinking of a
payment which already generated an account move.

However, no 'Cancelled' state has been introduced for the payments. It
means that the cancellation of a payment will set it to a draft state,
and it won't be possible to delete or flag it as 'Cancelled' later on.
It will therefore be a misleading entry for the users.

opw-694351