[IMP] web_editor: ace_editor: right resize bar
Before this commit, it's was only possible to resize the width
of the ace editor when it was on the right side of the screen.

In studio, we need to put the ace editor on the left.
So, we needed to add a new resize bar.
For that, we also add a new option to know the position of the
ace_editor.

We also prevent the width of the ace editor from being recorded
too many times in the local storage for nothing.
This is why we added _.debounce.