[FIX] project_issue_sheet: copy issue
- Create a project which handles Tasks and Issues
- Create an issue
- Book some timesheet on the issue, thanks to the tab "Timesheet"
- Go to the Timesheet application, you'll see that the time booked
  appears in the grid
- Change the time spent from the grid view.

The time added is booked on the issue, while the user doesn't expect it.

This is because the modification of a timesheet line copies an existing
line, then change the necessary values (see method `adjust_grid`).

Ideally, we should have a link module between timesheet grid and project
issue. Another solution would be some hack in the `timesheet_grid`
module to prevent `issue_id` to be copied. However, there is little to
no use of copying a timesheet line in general, so the solution chosen
should not affect the existing processes.

opw-697018