[FIX] mail: escape canned responses
Create a canned response containing the character "<". It is
automatically converted into "&lt;".

The substitution field is systematically escaped, for security reason.
This causes the substitution observed. However, the canned responses are
only used in text-only messages. The escaping is therefore not
necessary, since the HTML is not interpreted.

opw-693970