[FIX] website_sale: 2 state tags and same shipping change
If website_sale_delivery is installed, when we choose same invoice
address as shipping or another one, a new set of country/state are
displayed.

It is done since if we are using the invoicing address as shipping, the
address must be in the allowed shipping countries and states.

This was solved for some instances with 2f49f68d07f9 but failed in other
instances.

With this fix the two state select tags have references to their own
options which allow us to have it alright for the several possibilities.

opw-693127