[FIX] account: reintroduce tax lines in invoice
Commit 873da3c6215bb introduced the tax group feature. The tax line
details were removed from the printed invoice in favor of the tax group.

However, it requires that all tax groups are correctly defined in the
localization, otherwise the printed invoice doesn't show the required
tax details.

The correct fix, which is the complete definition of the tax groups,
will be done for the upcoming v11 version. In the meantime, this simply
introduces back the tax details table.