[FIX] product_extended: Compute from BoM
The price suggested thanks to the "Compute from BoM" button only takes
into account the price of the products listed in the BoM lines.

In v9, there was a "Recursive" option that provided the possibility to
compute the prices of all products implied in the BoM, recursively. This
option has been removed since it was leading to a "black box" behavior
that was unclear to the users.

However, limiting the price computation to the BoM lines only is too
restrictive since it doesn't even take into account the phantom BoMs.
By using the `explode` method, we have access to:
- all products not manufactured
- all products manufactured which are not kits (phantom BoM)

It means that the cost of manufactured products (not kits) should still
be computed before the product itself.

opw-697866