[FIX] web: document domain selector options correctly
The "ModelFieldSelector" widget has a "filters" option which can
customize the fields which are displayed.

The "DomainSelector" widget has of course the same option it can
transmit to its internal "ModelFieldSelector" instance... but it
was documented as "fs_filters" instead of "filters".

Also, let the "ModelFieldSelector" document its own options and
sets their defaults (instead of the "DomainSelector" widget).