[FIX] point_of_sale: avoid scrollbar on main window
Scrollbar appeared in Firefox (at least 50.1.0 and up). Related to
depending on non-standard behavior when setting height on a table. For a
detailed explanation see 0962886821005d6ffe5782d25594a5f87df89c70.

In this case we can simply ignore overflow on the main POS window since
the height difference introduced is only a couple of pixels.