[FIX] stock: correctly compute custom search method
The search on qty_available was refactored twice:
- 8eb1ce5c to split the method
- cba1b55d to remove the eval operator

At 8eb1ce5c the order of evaluation introduced a bug so that the domain
   ('qty_available', '=' 10.0)
was evaluated
   if product[qty_available] = 10.0:
which crashed (`=` instead of `==`).

Before 8eb1ce5c, this code did not crash as the operator was converted via
  if operator == '=':
    operator = '=='
which is no longer executed after the split

cba1b55d replaced the eval by a dictionnary that replaced the eval error by a
KeyError when searching with `=` operator.

This commit changes several things:
- replace the `operator in ('==', '>=', '<=')` as this was evaluated before the
  operator replacement code mentionned above
- remove the replacement code as no longer justified since cba1b55d
- replace the OPERATOR key `==` to always match given key