[FIX] web: do not call session load_qweb on frontend
Since commits 672dfec13db066970f97092544d0829c0c094e31 and 1ef0d815397e5cc6e38c5488407e2f57154eba02,
the session is not started by the webclient anymore but directly by the
session service instead, and the session info is found in the DOM
instead of by RPC.

Because of that, the session is started in the frontend and other
non-backend environments which caused a JS error while trying to
load qweb (as no qweb has to be loaded). This commit makes the
session start check if it is a non-backend environment and only
initialize session info (from DOM) in that case.