[FIX] account: register payment with payment, invoice and company having different currencies
In revision c0602d12c5917fab26c375acc381d910ac4559c6,
`_compute_total_invoices_amount()` was assumed to
give the invoice residual within the company currency but,
in the case the payment_currency is different than the company_currency,
it returns the amount in the payment currency.

In this case, we actually need the total invoices residual
within the company currency.