[FIX] web_kanban: load grouped field if not in view
In the kanban view, if we try to group by a field that is not
defined in the arch, some features (like 'Add a new Column')
are not available because the case is not correctly handled.

To fix it, a RPC is done to load all fields in this case.