[FIX] sale: Traceback product_packaging on SO line
When we add a SO line on a sale order with a product packaging, a
traceback is raised.

The problem comes from a default value and a domain set in the context
which reference a product template but since this rev: 271a67bcd79de4bc7a737d5feba7fb1c0886abcd
the product packagings are linked to product_product and not
product_template anymore.

So we used the product_id field instead of product_tmpl_id. And remove
product_tmpl_id from the sale_order_line model which is not used anymore