[FIX] base: company contact data
As a portal user, access one of your invoices from the frontend
(download the PDF). The generation of the report crashes because of an
access error. The error arises when trying to access the company contact
information (address, phone, fax...) displayed in the report footer.

This is due an API migration error (967ee94426e). Before the migration,
the partner was browsed in `_get_address_data` as:
`address = part_obj.read(cr, openerp.SUPERUSER_ID, ...`

After the migration, this is not browsed as superuser anymore.
Therefore, users with limited access rights won't be able to read it
anymore.

opw-708581