[FIX] stock_account: costing method on product category
The field "costing method" on the product.category only appears when
the inventory valuation is set to "perpetual (automated)" which is
wrong. The user must be able to choose the costing method no matter
which inventory valuation method is chosen.

The problem is due to this commit 70ebcc2cb7d34f087bf9f5c070222cc05d57840f
that changes the xml view and make the field invisible if the inventory
valuation is set to "Periodic (manual)".

This commit revert the change on the xml view and thus add again
the group attribute that should not have been removed.