[FIX] hr_timesheet: computation of the children hours according to the child stage
The below revision:
63dc8546ec09ac736937c0d957e71b4ebc98b6c5

Change the behavior on how are computed the children hours
according to their stage fold state, while it was not intended
to. The subject of the above revision was an issue
in the recomputation dependence for the field
`children_hours`, but not how the field was computed.

It's important to count the effective hours when the
sub task is in a final stage (e.g. done) rather
than the planned hours in this case.

opw-709008