[IMP] delivery,product,sale_stock,stock: split packaging views
As the product.packaging model is used for two different business cases,
the views shown when the model is used for product packaging must only
contains a product_id, barcode, and qty because and when it's used for
delivery packages we must show the carrier, barcode, size and wheight.

So for usability purpose, we have made two different views explicitly
called in each use case.