[FIX] account: prevent loops when unreconciling entries
13e93e95a037d2546c7e0ae94fd18c8b770a9662 refactored the handling of
exchange rate entries when unreconciling entries.

However it did not preserve the piece of code that prevented loops
during this operation (`full_rec_lookup` context key).

This could cause the unreconciliation of entries with an exchange rate
difference to fail with a confusing error claiming that the user is
trying "to reconcile some entries that are already reconciled!".
The double reconciliation was caused by a recursive attempt to reverse
the exchange rate entry.