[FIX] web_kanban: use string when group by selection
In the kanban view, if we try to group by a selection field that is
not defined in the arch, the key will be used as column name (instead
of the string).

A fix has already been pushed to load all the fields in this case
(see 8ec950a) but these fields were not correctly used in this
particular case.