[FIX] product: get_names with empty record set
Steps to reproduce:

-Go to accounting dashboard
-Click on "More" in vendor bills dashboard
-Click on "bills analysis"

Bug: it opened the account.invoice.report view, which crashed.

Reason: The function get_names in addons/product/models/product.py got an empty record set
product.category(None,) and returned [False] so when joining " \ " with [False], it crashed.
The empty record set was returned  when the categ_id of an account.invoice.report() was None.

opw:724783