[FIX] web_calendar: timezone on event display regression
In babd9139 event that were not displayed for arabic were fixed but
there could be an error when displaying timezone caused by this change.

As said in https://fullcalendar.io/docs1/event_data/ignoreTimezone/ the
ignoreTimezone parameter is set to `true` by default in FullCalendar 1
(this is not the case in current FullCalendar).

So there is two way to solves this time zone issue:

- keep babd9139 and have the parameter ignoreTimezone set to `false`
- use momentjs `.toString()` to get a javascript english date string
  which will thereafter be parsed by javascript `new Date()`.

This commit use the second way since it is nearer what was originaly and
hopefully safer.

opw-705752

momentjs doc on toString: http://momentjs.com/docs/#/displaying/as-string/