[FIX] account: usability of entry sequences.
This revert commits 866f0f3b030656cb8acfb35c90e639f7d8b9629b and a7ae991a73ff3f74091a55fb4f20ca2c84e68e42 that were introduced to solve OPW-693955.

The usability before those fixes was good, but the bug was appearing when we first create the journal without the boolean 'dedicated refund sequence' set, then we try to modify that va$

This time, the fix consist to create automatically the sequence on the write of the boolean if it's not existing yet, in the same fashion than on the create().