[FIX] account: fix statement_line_id on bank statement processing.
Since 0be9eacf3325e54e4452ef5169ed80a263046eb3, a new field 'statement_line_id' has been added on account.move.line object and the field 'statement_id' became a related... but this statement_line_id on the bank account aml was, by mistake, not set. That created aml that were wrongly appearing in blue in the reconciliation widget.

If this error occured in a DB, the following SQL query can fix it:
$ UPDATE account_move_line l SET statement_line_id = s.id FROM account_bank_statement_line s
   WHERE l.statement_line_id IS NULL
   AND l.statement_id = s.statement_id
   AND l.name = s.name
   AND coalesce(s.partner_id, 0) = coalesce(l.partner_id, 0);