[FIX] mrp: BoM count detection in case it is a single variant
In case a product is a single variant we should not leave it with a BoM
of zero as soon there is a BoM for its template. This would for example
hide the `compute_price` object button which relies on the `bom_count`to
be visible.

BoM is not found for a single variant product and compute price not
possible from the `product.product` view although it is implemented
actually there.

Closes #16237