[FIX] *: unsupported type for graph views
In the new implementation, the type is checked to prevent loading
a view type that does not exist.

In the old implementation, the type was ignored resulting in a graph view.

So, before this commit, some views were no longer loaded.

The 'pivot' type should not be used since there is a dedicated pivot
view.