[IMP] maintenance : responsible added as follower
This commit add the following feature :
Add the responsible and the owner of the maintenance request as
follower.

This commit creates a new method that adds as follower the responsible and/or
the owner. This method is triggered in the create and in the write if they
are modified.

It will not remove old responsible or owner from the chatter.