[MOV][ADD] website_account: move account code out of website_portal_sale
Purpose: separate and locate website code for main addons in their
correct module. This commit targets accounting related customer portal.

This commit adds a new bridge module between website and account. It
contains code related to website portal of invoices. It was located
in website_portal_sale but this one should only be the quotation and
sales portal.

A dependency from website_sale to website_account is added to ensure the
behavior and available features of eCommerce are not impacted by this
commit.

No code modification is performed in this commit except one unnecessary
less rule. It contains only code move.