[FIX] product,stock: create a company for non-admin users
The typical multi-company record rule is
['|',('company_id','child_of',[user.company_id.id]),('company_id','=',False)]

As the company is just created, the current user does not belong to the company.
Create first without company (to still check the access rights) and after add
the company in sudo

Can not apply the same patch for stock as locations and sublocations are created
based on the warehouse company.

Fixes #16807