[FIX] website_form: do not subscribe administrator to all records from website form
As administrator is used to create all records from website form he is
also put into followers. This creates a lot of unnecessary notifications
and/or emails depending on the system configuration.

Using the magic context key this behavior is modified. Administrator
will not follow every records created through the website form anymore.