[FIX] product: non-random default variant attribute order
Before this revision, if the user did not re-order
the sequence of its attribute values manually,
the attributes in the variant name were ordered
randomly.
e.g. On the ipod, one time it could be the color before
the GB, one time the other way around.

Besides, this could lead to an issue in the pivot table,
as mentioned in the revision
f343715123bb36d33782c98e9dd0149f29567115.

However, on the contrary on what is done in the above revision,
these attributes must not be ordered on their value name,
but rather on their sequence, and if they have the same,
on their attribute, so the order is always deterministic.