[FIX] point_of_sale: don't reserve pickings with tracked products
Introduced by 9f03871f221766a5409255a6bdc76003634386c0 which was
intended for 9.0. It does not work well in 10.0 because here users
have the possibility to specify their own tracking numbers directly in
the POS interface. This causes issues when eg. the user specifies the
same lot number as is already reserved by the system.

opw-742657