[FIX] web_editor: don't resize iframe in content
In 566f5a484 an iframe for the readonly preview was added when
displaying html fields.

In b514ac54c this iframe was only used if style-inline was set in the
widget options.

This brought an incompatibility if the not iframed preview contained an
iframe (eg. a youtube video which is then resize to a 30px height) which
is solved by this commit.

opw-725161
closes #16955