[FIX] event_sale: auto_confirm free ticket
After a7c08218 a registration was confirmed once an invoice was fully
paid.

This is not very handy if the registration was free to begin with, you
need to create a 0.00 amount invoice.

This commit get back the previous behavior of auto-confirming
regitration directly when the registration is free.

opw-742901
closes #17035