[FIX] stock_account: inventory valuation: default value
... for 'compute_at_date' field (default 0, 'Current Inventory').

Before the new views, this value was selected by default by the
FieldRadio, but it was a mere coincidence. Indeed, as the field
had no default value, no value was specified in the default_get,
and for no reason, the widget selected 0 as value.

Anyway, as this behavior was wanted, we simply specify a default
value in the action.