[FIX] crm: quick create in crm.lead
When making a quick create for a crm.lead record, the field user_id is set
with its default value (introduced by d566c1fbd44409d82e99f81ee75e8227503b26ba)
But the date_open('Assigned') was not set.

When creating a challenge with goal="New opportunities", the new opportinities
are counted with date_open. But when creating an opportunity with the quick create,
it was not counted as date_open was empty.

opw:742751