[FIX] point_of_sale: fix forward-port of d7c5101d
Introduced at 6a375f21d06b54c71311a216afd9b4f10a6fc12b. Normally not a
huge issue (you just end up with a '>' character). But unfortunately
in this case it crashes the hw_escpos driver since it tries to parse
the content of that element as a float.

opw-744222