[FIX] mrp_repair, sale: display the right partner.vat
Some reports displays the wrong partner.vat.
On the invoice report, when the shipping address is different than the
invoicing address, we also also display the shipping address in addition
to the invoicing address. But we displayed the VAT of the invoicing
address instead of the one from the shipping address.

Same issue on the MRP repair report and the sale report.