[FIX] account: revert a49abdbdb and proper fix
Revert "[FIX] account: reset the move reference on bank statement"

This reverts commit a49abdbdbad61f09040e51b50fb6bd49c75dd0a7.

The previous patch was removing the reference of the move.
However, the reference is useful to reuse the name if the line is reconciled
again.

Instead, remove the constrain. To delete the move, account_cancel is already
needed. It is useful to allow removing statement lines to match the statements
form the bank and the one on Odoo.
In the future, allowing to set the amounts to zero should be added as a way to
correct the lines.

opw-741205