[FIX] website_sale(_delivery): limit 'Setup payment / delivery methods' to group_system
When choosing a payment acquirer or a delivery provider there is a link
allowing to directly jump in the backend to configure them. It has been
introduced at c30818c44a3bda62fb29060604b3de5d365b8a16.

However it is displayed for every user going through the checkout process.
This commit limits it to group_system. It seems a better idea to not show
it to everyone as they don't have access to it anyway.