[FIX] account: children in Invoiced partner button
When we are on a partner form view, we get the total invoiced containing
the child contact of the current partner. But when clicking the button
we only see the invoices of exactly the current partner.

In most locations of the code this has changed (since 347c784da) and
this part should also contain the descendants contacts.

opw-744929
closes #17188