[FIX] web: pivot view should use col_group_by and measures key
For historical reasons, when the pivot view was rewritten, we lost the ability to understand the measure and col_group_by keys from the context.

It was actually a deliberate decision, because we decided that it was not possible to properly handle the colgroupbys in the search views, so we replaced that mechanism with pivot_col_group_by and pivot_measures keys exported by the view, when necessary.

Sadly, we did not properly remove the outdated keys from the xml files.  So, with this commit, we reintroduce those special keys, because 10.0 is a stable version.  Note that this will be removed in master.

 (PR 17119)