[FIX] web_editor, website: allow to detect indirect clones
When cloning a snippet section thanks to the editor, the associated
snippet options' `on_clone` methods are called. Now these methods
receive a new argument `options`. The option this commit introduces
is the `isCurrent` one. It allows the method to know if the snippet
section it is associated to is the one that was cloned or if it is a
child of the one that was cloned.

This is useful to solve a known bug when cloning elements which have
bootstrap columns:

- When one column is cloned we want it to lose its col-*-offset-*
  classes so that the offset is not duplicated (only the column).

- When a container with column children is cloned we do not want
  these column' offsets to be removed (we want the whole cloned
  container to be the same).