[FIX] point_of_sale: avoid '414 Request-URI Too Large' on config form
Inspired by aa20cf03e34a90211c6122b15aa117258d5741ca. Without this you
end up with a URI with a length of over 50000 characters. The default
template doesn't actually use any fields on pos.config so we could
theoretically get rid of everything. But because it's easy enough we
just get rid of relational fields and the customer_facing_display_html
field itself which make up the vast majority of the URI.