[FIX] project: merge tasks should be filled automatically.
With the new views, the merge task feature did not properly precomplete the selected tasks.  This is because they do not support the command 4 when creating a new record.  This will be fixed at some point, but meanwhile, we can just simply return the selected ids in the default_get.  This makes the code slightly simpler.

(PR #17566, issue 746867)