[FIX] point_of_sale: fix included taxes
Really apply commit 4433f29ff12 to POS orders.

Since the creation of a POS order in the backend is a feature which
nearly not used, we rewrite everything related to onchanges. To avoid
breaking the compatibility, the old onchange code is kept. As long as
the user doesn't upgrade the module it will be used instead of the new
onchange.

opw-746827