[FIX] web_editor: style if website not installed
web_editor can be installed without website, thus missing needed CSS
causing e.g the widget sidebar to overlay and hide the mass mailing
content if website is not installed.

This fix implement a simple middle ugly solutions, but doing something
more in stable would be an hassle (for example in bc00ceff the #wrapwrap
CSS is included in report module).

opw-748975
closes #18005