[FIX] website_sale: Sort by catalog price in the shop
When sorting by price (increasing or decreasing) in the shop,
the products were sorted according to the catalog price(list_price)
and not according to the price(which takes the pricelist into account).
So a better name is given in this patch.

Reason: The field "price" is a computed and not stored field.
Then it's not possible to order the products with it in the orm.

opw:750997