[FIX] web: avoid interactions between drag and drop tests
The commit 97bc95467bc5ab422ebb66b5d79431a52a9a99b4 introduces a new option
to drag element without dropping them.

But it's not allowed to start a new drag and drop until the first one is
completely done.

So when the element is removed from the dom (the test is finished), we complete
the drag and drop in order to let next tests perform drag and drop.