[IMP] account: allow to switch accounting packages (CoA...)
We can now select a new localization module in place of the existing one, if no accounting has been done yet.

This allows to set a default CoA for Germany and China, by choosing arbitrary one of the existing localization package at the installation since they can directly switch if needed.

Was PR #16605.