[FIX] stock: add move when picking is available
This completes rev e0a8fac2ec0897d937a3b6689143cd94adb2154d.

The state field was correctly updated by the default_get, however it was
not set to the server because it was set as readonly on the model. For
this specific usecase, we force the readonly to False in the view.

The same issue was happening with the product_uom_id, which is set to
readonly in the view. We mark the field as editable if the move is
additional, and as it does not reserve anything it shouldn't trigger any
other issues.