[FIX] stock: improve warning to help user to find blocking move.
Add into the warning message the list of stock.move that raise the constraints
error.

When you create an 'Inventory Ajustment' of 1000 lines and see this error,
it is not easy to find which product could cause the problem.

This solution is not perfect since the name is not uniq, but it decrease
drastically the number of record to check manually.

opw-760643