[FIX] mail: systray: next activity not closed on click
Before this commit, when you clicked on the dropdown, the action
was executed but the dropdown wasn't closed.
This issue was very annoying on mobile because the dropdown is in fullscreen
and we couldn't see the result of the action.

Now, we removed the "stopPropagation" which stops the bootstrap event
to close this dropdown.