[FIX] crm: correct activities filter combination
The new "My Activites" filter introduced by rev. c57ac45d78bcfc300b781691fb4a5da6c7d81c7a
should not be in the same group of the search view as other activity
filters (Overdue/Today/Upcoming), otherwise they are combined with OR
instead of AND.

This is particularly misleading when coming from the Sales dashboard, as
the "Overdue" button of "My Pipeline" will lead to a list of
opportunities with "My Activities OR Overdue Activities", showing *all*
opps with overdue activities, not just yours.