[FIX] web: correct first render of calendar when changing the date
As mentioned in
https://github.com/odoo/odoo/commit/2bf1b0620fa969382b89daa12fc31d88202f40ef,
a resize after the first render is necessary to prevent overflow issues
with fullCalendar. As changing the date would rerender the calendar, we
also need to trigger a resize then.

From this revision onwards, a resize event is triggered every time the
calendar has finished rendering its events.

Unfortunately, I was not able to write a meaningful and successful test
(I encountered PhantomJS-specific issues). I also removed the test for
https://github.com/odoo/odoo/commit/2bf1b0620fa969382b89daa12fc31d88202f40ef
because it always passed regardless of whether the fix is applied.