[FIX] portal: fix frontend chatter move
Commit 038f7ce7750b6a7a65ee70fa2b6b58cf7d163b7b moved the frontend chatter from a website-based module
website_mail to portal. Purpose was to allow its use in customer
portal independently from the website.

However some renaming done during the move was not properly propagated
resulting in the chatter not being loaded. This commit fixes that by
correctly using the new class node as well as widget name when
instantiating the widget.