[FIX] auth_signup: fix computation of default template user in settings
Indeed we store an integer that is the ID of the user to copy when
creating the portal user, not a boolean. We therefore have to evaluate
the result instead of comparing it to a boolean. This was introduced
at f18c323cba2f41a04e6392f8727c2c256a74f0d2.