[FIX] stock: calendar view should not be removed
I deleted calendar view from picking because I thought it was not
used anymore. However it was used in the menu item action thus it
returns a traceback when used and the picking menu was not accessible
anymore.

This commit revert commit 296dfc14e65a501726fc2623b56fbe376cb10014
and fix the view instead by using schedule date instead of min_date