[FIX] web_editor, website_sale: load images of the shop when product has been updated
Before this commit, when the shop was in multi images, the main image was not refreshed when changed in the back end.
This was because of the ?unique parameter that was computed on the product variant image. Hence, when the product image changed, the variant image did not, the hash neither, and the image wasn't refreshed

This commit fixes this by adding an option to the qweb.field.image

OPW 765287

closes #19133