[FIX] hr_timesheet: security, do not require all timesheets user to be HR officer.
Regular employees are added to the timesheets user group
so they can access to the Timesheets modules.

Requiring timesheets users to be HR Officer would almost mean to
require all employees to be HR officer, which is not the case.

The above is the result of the revision
c7960a9abc0d22aeee0421cfe0ab845dd7d355b8

Before, the timesheets user group was the
timesheets officer group, and it was the only
group for the module timesheets. There
were no timesheets manager or user group.
This group implied the HR Officer groups,
but it made sense as it was the Timesheets
Officer group.

Now that it is the timesheets user group,
it doesn't make sense to imply
the HR Officer group.