[FIX] sale: fix defaults taxes on tests in sale
Suppose two ways to install sale:
-i account followed by -i sale: account is installed, the execute is done on the chart template and then,
sale is installed. In this case, the IrDefault values already exists when sale is loaded.

-i sale directly: account and sale are installed, the execute is done on the chart template.
The difference is that there is no IrDefault values set on the demo data in sale.

The problem is that such behavior can lead to wrong tests. For exemple, having taxes_id/supplier_taxes_id
set or not set on product.template during sale tests.