[IMP] sale_payment: support payment in sale customer portal
This commit adds support of payment in the sale customer portal. This
is done in the sale_payment module and uses the settings introduced
previously about online confirmation and payment. This way customers can
now pay orders without having to rely on online quote mechanism.

Two routes are defined in this commit, one for form-based payments and
one for server2server-based payment. Please refer to commit 602c27bedbe13180253b2858c0ad0431e380608a
for more details about the new payment form.

Some small modifications are done SO confirmation and invoice generation
methods to return error codes that can be used in the customer portal.
As return values were not used in website_sale or website_quote this
should not impact those modules.