[IMP] sale_stock: raise multiple error at the same time
When a user confirm a sale order, he will have one error at a time.
This is not user friendly since he has to correct the current error
and confirm the sale order again in order to get the next mistake.

This commit runs all procurement and catch raised error in order
to display them all at once