[FIX] sale_stock: partner not propagate to picking.
Since procurement refactoring in commit 64af85942d96c77b8b9caa6f3618e17274a9e20e
the partner was not propagate between sale order and picking.

It happens because the _run_move method uses the group's
partner on the picking but the partner was not set on the
group during its creation.

This commit add the shipping partner id of the order
in the create method and this partner is used later in the
picking/move create method.