[FIX] stock: priority on new move compare bool and str
Since python3 compare boolean and string is not
possible anymore. Priority field on stock move line is
not required and thus it happens that we compare a set
priority with an unset one.

In order to correct it this commit only compares move line
that have priority and if tey are not use the default priority
1.