[FIX] sale: add name_get for sale order line
Since procurement is dead, selecting a SO line
on task to reinvoice it get the list of so line
description in dropdown.
Define a 'name_get' method is required. So avoid
performance issue, we only customize it (order name
and product name) with a context key, like we did
for Sales Order.