[FIX] web: focus issues after onchange in x2many with buttons
https://github.com/odoo/odoo/commit/2c00d1e824a9920225ec1288c165a63e61a1d0b2
fixes most post-onchange focus issues for x2many fields, but does not
account for "button columns" properly. More precisely, the current
widget is computed via a position index taking buttons into account in
an array that ignores them. This caused the index to be out of bounds,
resulting in a traceback. This commit fixes this issue.