[FIX] mail:  thread note bg-color
This commit fix 49f9360 , using a better bg-color in order to improve
text readability. It apply also apply a minor padding-left to avoid
having text attached to the coloured div border.
Moreover, the bg-colour can not be applied to activities thread.

The little history behind: remove the bg-color was decided by fp and
the UX dep. I’m not aware if 49f9360 is approved or not, but the final
result was not acceptable.