[FIX] web: Kanban: open record when a tag is clicked
Kanban records with 'global click' feature activated should open
the record in a form view when the corresponding card is clicked.
Before this rev., it wasn't working if the user clicked on a tag
of the kanban record, because of an unused click handler defined
on the Many2ManyTags widget (preventing the global click feature
from actually performing the global click).

This rev. removes that handler in the KanbanFieldMany2ManyTags.