[FIX] mrp: make sure the note is logged when you change done qties
The production_id on the stock move line is a helper field for work
orders and you can not assume it is always set correctly.  As
non-tracked stock moves create the stock move lines themselves it is not
set.

So, for logging notes, we check the 2 production order fields through
the moves.