[FIX] purchase: amount of the purchase in the logged message
When a picking is created from a purchase order, a message is logged in
the chatter:  "This transfer has been created from: POXXX: XX$".

Ideally, an inventory user should not have access to the purchase
orders. Sadly, removing the access right brings a lot of access right
issues, so this commit only remove the amount from the logged message

We adapt the name_get that always displayed the total amount of the
purchase order since rev[1], as a message was logged in the chatter when
a picking is created from a purchase order and the message contained this
value.

We implement a context key to display the total amount in the name_get,
as it is used when linking a vendor bill to a purchase order.

[1] 83bd339fa81d6258b0401963ab1d60260ca73692