[FIX] point_of_sale: rounding error when create picking
When a return picking is created from a POS order, the methods check
that the value rounded to the unit of measure is not null.

The problem is that a rounding is passed to the function 'float_is_zero'
in the parameter precision_digits instead of precision rounding.

This bug has been introduced in rev: https://github.com/odoo/odoo/commit/0d51b507dc7343fe492804b8a71da7a337bcf1b1

opw-771282